[spaces] moved RaviartThomas -> RT, refs #14
